♻️ Refactor: API Endpoint 변경 (/api/analysis/stores/{store_id})
