+++ app/schemas/analysis_schema.py
@@ -3,6 +3,6 @@ from pydantic import BaseModel, Field
 from app.schemas.action_plan_schema import ActionPlanResponse
 
 
-class StoreAnalysisResponse(BaseModel):
+class AnalysisStoreResponse(BaseModel):
     swot: List[str] # 임시
     action_plan: ActionPlanResponse
