+++ app/services/analysis_service.py
